Bug 34390: Fix inconsistencies in credit types titles, breadcrumbs, and header
Fix inconsistencies in credit types titles, breadcrumbs, and header

This patch fixes a couple of inconsistencies in the credit types
administration page, making sure the page title, breadcrumb
navigation, and page headers are consistent with each other.

To test, apply the patch and go to Administration -> Credit types. Test
each variation of the page:

- Main page
- New credit type
- Modify credit type

Signed-off-by: David Nind <david@davidnind.com>
Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>
Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>