Bug 29002: Fix item select options
SQUASH INTO RESET BUG?

This patch ensures that when a period is not selected we remove all item
selection limitations from the item select.

Prior to this patch, if you clicked to edit a booking, cancelled, then
clicked to place a new booking.. the modal would prevent selection of
the item which you just edited.

Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>
Signed-off-by: Janet McGowan <janet.mcgowan@ptfs-europe.com>
Signed-off-by: Caroline Cyr La Rose <caroline.cyr-la-rose@inlibro.com>
Signed-off-by: Laurence Rault <laurence.rault@biblibre.com>
Signed-off-by: Kyle M Hall <kyle@bywatersolutions.com>
Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>